Rearrange server impl
